Generators: put files into work_root instead of cache_root
Generators currently place generated files in cache_root, which is
typically ~/.cache/fusesoc/generated. No locking is used when creating
files in this shared folder.

This behavior leads to race conditions when multiple parallel fusesoc
runs try to read/write files from this directory, leading to e.g. the
case where a generator in process 1 reads a half-written file from a
generator in process 2.

To avoid this behavior, the generator-produced output files are now put
into a subdirectory of work_root. This is consistent with other
fusesoc-generated files, which are also put into work_root.

To avoid picking up these generated core files unintentioanlly, e.g.
because there's an older build folder in the source tree, a
FUSESOC_IGNORE file is added to work_root. The library of generated core
files is explicitly added.